+++ src/hooks/LevelBrowserLayer.cpp
@@ -134,8 +134,8 @@ void LevelBrowserLayerHook::onLogin(Result<std::string> result) {
 	// });
 
 	// auto req = WebManager::get()->createAuthenticatedRequest();
-	// req.param("old_account_name", "username1");
-	// req.param("new_account_name", "username2");
+	// req.param("old_account_name", "Pugs20");
+	// req.param("new_account_name", "PuggyBuggys");
 	// m_fields->adminTask.spawn(req.post(WebManager::get()->getServerURL("admin/transfer_account")), [this](auto response) {
 	// 	auto result = response.string();
 	// 	if (result.isErr()) {